requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
pydub>=0.25.1
yandex-speechkit>=1.3.0 
yandexgpt-python>=0.2.3
asyncio>=3.4.3